added collab server id to stars, say command now is randomized

+++ src/commands/utility.ts
@@ -10,12 +10,30 @@ export async function pingCmd(client: Client, args: string[], message: Message) 
 }
 
 export async function sayCmd(client: Client, args: string[], message: Message) {
+    const sayMessage:string  = args.join(" ");
+    const userId = `<@${message.author.id}>`;
+    const deniedMsgs = [
+        `Sorry ${userId}, there\'s a 5% chance i\'ll actually say that.`,
+        'Reh',
+        ':RioluUgh:',
+        `I\'m gonna send you to the ranch, ${userId}`,
+        `You have no power over me, ${userId}`,
+        `I\'m not gonna say that ${userId}...`,
+        `Why should I say ${sayMessage}?`,
+        `${sayMessage}???????`,
+        `${args.reverse().join(" ")}`,
+        `${sayMessage.toUpperCase().substring(0, Math.round(sayMessage.length / 2))}-`];
+    if (!isAdmin(message) || Math.random() < .95) {
+        await message.channel.send(deniedMsgs[Math.round(Math.random() * deniedMsgs.length)]);
+        return;
+    }
     // makes the bot say something and delete the message. As an example, it's open to anyone to use.
     // To get the "message" itself we join the `args` back into a string with spaces:
-    await message.delete();
-    const sayMessage = args.join(" ");
+
     // And we get the bot to say the thing:
     await message.channel.send(sayMessage.length ? sayMessage : `You didn't say anything! >:(`);
+    await message.delete();
+
 }
 
 export async function helpCmd(client: Client, args: string[], message: Message) {
